refactor(OUT-1478) : applies requested changes, separation of activityWrapper optimistic update methods into a new util
