+++ src/components/layouts/ConfirmDeleteUI.tsx
@@ -24,7 +24,7 @@ export const ConfirmDeleteUI = ({
   const getHeaderMessage = () => {
     if (customBody) return customBody
     if (bodyTag === 'comment') return 'Delete comment?'
-    if (bodyTag === 'message') return 'Delete message'
+    if (bodyTag === 'message') return 'Delete message?'
     return `Are you sure you want to delete this ${bodyTag}?`
   }
